Idea para crear el ejecutable.
