+++ Resources/.env.example
@@ -1,0 +1,1 @@
+API_KEY=
