+++ Resources/Pipfile
@@ -1,0 +1,12 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+flet = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.11"
