--- main.py
@@ -1,6 +1,0 @@
-#importar cache
-#importar gui
-
-# inicializar cache
-
-# inicializar gui
